# be/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUseCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["user"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["user"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"]},"put":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUserUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["user"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"]}},"/user-career":{"post":{"operationId":"UserCareerController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUserCareerCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["user-career"]},"get":{"operationId":"UserCareerController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"default":1,"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"default":10,"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user-career"]}},"/user-career/{id}":{"get":{"operationId":"UserCareerController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user-career"]},"put":{"operationId":"UserCareerController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUserCareerUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["user-career"]},"delete":{"operationId":"UserCareerController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user-career"]}},"/promo":{"post":{"operationId":"PromoController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"Promo Spesial Akhir Tahun"},"subtitle":{"type":"string","example":"Diskon hingga 50% untuk semua produk"},"description":{"type":"string","example":"Nikmati potongan harga besar untuk berbagai produk pilihan selama periode promo."},"syaratKetentuan":{"type":"string","example":"Berlaku untuk pembelian minimal Rp500.000 dan tidak dapat digabung dengan promo lain."},"berlakuHingga":{"type":"string","example":"2024-12-31"},"brandId":{"type":"string","format":"uuid"},"image":{"type":"string","format":"binary"}},"required":["title","subtitle","description","syaratKetentuan","berlakuHingga","brandId","image"]}}}},"responses":{"201":{"description":""}},"tags":["promo"]},"get":{"operationId":"PromoController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"type":"string"}},{"name":"brandId","required":false,"in":"query","description":"Filter by brand ID","schema":{"example":"9b62c4d3-5f25-4f8f-83cf-3f49a6b8fd7c","type":"string"}}],"responses":{"200":{"description":""}},"tags":["promo"]}},"/promo/{id}":{"get":{"operationId":"PromoController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["promo"]},"put":{"operationId":"PromoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"Promo Spesial Akhir Tahun"},"subtitle":{"type":"string","example":"Diskon hingga 50% untuk semua produk"},"description":{"type":"string","example":"Nikmati potongan harga besar untuk berbagai produk pilihan selama periode promo."},"syaratKetentuan":{"type":"string","example":"Berlaku untuk pembelian minimal Rp500.000 dan tidak dapat digabung dengan promo lain."},"berlakuHingga":{"type":"string","example":"2024-12-31"},"brandId":{"type":"string","format":"uuid"},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"tags":["promo"]},"delete":{"operationId":"PromoController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["promo"]}},"/brand":{"post":{"operationId":"BrandController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"nama":{"type":"string","example":"Bakso Enggal"},"description":{"type":"string","example":"Restoran bakso prasmanan pertama di Indonesia."},"logo":{"type":"string","format":"binary"}},"required":["nama","description","logo"]}}}},"responses":{"201":{"description":""}},"tags":["brand"]},"get":{"operationId":"BrandController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["brand"]}},"/brand/{id}":{"get":{"operationId":"BrandController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["brand"]},"put":{"operationId":"BrandController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"nama":{"type":"string","example":"Bakso Enggal"},"description":{"type":"string","example":"Restoran bakso prasmanan pertama di Indonesia."},"logo":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"tags":["brand"]},"delete":{"operationId":"BrandController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["brand"]}},"/outlet":{"post":{"operationId":"OutletController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOutletCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["outlet"]},"get":{"operationId":"OutletController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"type":"string"}},{"name":"brandId","required":false,"in":"query","description":"Filter by brand id","schema":{"example":"9b62c4d3-5f25-4f8f-83cf-3f49a6b8fd7c","type":"string"}}],"responses":{"200":{"description":""}},"tags":["outlet"]}},"/outlet/{id}":{"get":{"operationId":"OutletController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["outlet"]},"put":{"operationId":"OutletController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOutletUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["outlet"]},"delete":{"operationId":"OutletController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["outlet"]}},"/team":{"post":{"operationId":"TeamController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestTeamCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["team"]},"get":{"operationId":"TeamController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["team"]}},"/team/{id}":{"get":{"operationId":"TeamController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["team"]},"put":{"operationId":"TeamController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestTeamUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["team"]},"delete":{"operationId":"TeamController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["team"]}},"/upload":{"post":{"operationId":"UploadController_upload","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"201":{"description":""}},"tags":["upload"]}},"/upload/{filename}":{"get":{"operationId":"UploadController_serve","parameters":[{"name":"filename","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["upload"]}},"/berita":{"post":{"operationId":"BeritaController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"judul":{"type":"string","example":"Grand Opening Outlet Baru"},"createdDate":{"type":"string","example":"2025-01-15"},"penulis":{"type":"string","example":"Tim Marketing"},"content":{"type":"string","example":"Kami dengan bangga mengumumkan pembukaan outlet baru dengan berbagai promo menarik untuk pelanggan setia."},"image":{"type":"string","format":"binary"}},"required":["judul","createdDate","penulis","content"]}}}},"responses":{"201":{"description":""}},"tags":["berita"]},"get":{"operationId":"BeritaController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["berita"]}},"/berita/{id}":{"get":{"operationId":"BeritaController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["berita"]},"put":{"operationId":"BeritaController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"judul":{"type":"string","example":"Grand Opening Outlet Baru"},"createdDate":{"type":"string","example":"2025-01-15"},"penulis":{"type":"string","example":"Tim Marketing"},"content":{"type":"string","example":"Kami dengan bangga mengumumkan pembukaan outlet baru dengan berbagai promo menarik untuk pelanggan setia."},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"tags":["berita"]},"delete":{"operationId":"BeritaController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["berita"]}}},"info":{"title":"EnggalGroup API","description":"The EnggalGroup API description","version":"0.1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RequestUseCreateDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"nama":{"type":"string"}},"required":["email","password","nama"]},"RequestUserUpdateDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"nama":{"type":"string"}}},"RequestUserCareerCreateDto":{"type":"object","properties":{"tanggal":{"type":"string","example":"2024-05-20"},"nama":{"type":"string","example":"Budi Santoso"},"no_hp":{"type":"string","example":"081234567890"},"email":{"type":"string","example":"budi@example.com"},"cv_link":{"type":"string","example":"https://example.com/cv/budi.pdf"}},"required":["tanggal","nama","no_hp","email","cv_link"]},"RequestUserCareerUpdateDto":{"type":"object","properties":{"tanggal":{"type":"string","example":"2024-05-20"},"nama":{"type":"string","example":"Budi Santoso"},"no_hp":{"type":"string","example":"081234567890"},"email":{"type":"string","example":"budi@example.com"},"cv_link":{"type":"string","example":"https://example.com/cv/budi.pdf"}}},"RequestOutletCreateDto":{"type":"object","properties":{"kota":{"type":"string","example":"Jakarta"},"jamOperasional":{"type":"string","example":"08:00 - 22:00 WIB"},"lokasi":{"type":"string","example":"Jl. Jenderal Sudirman No. 1, Jakarta"},"image":{"type":"string","example":"/uploads/outlet-jakarta.png"},"googleMapsLink":{"type":"string","example":"https://maps.app.goo.gl/abcdefghijk"},"whatsappUrl":{"type":"string","example":"https://wa.me/6281234567890"},"brandId":{"type":"string","example":"9b62c4d3-5f25-4f8f-83cf-3f49a6b8fd7c"}},"required":["kota","jamOperasional","lokasi","image","googleMapsLink","brandId"]},"RequestOutletUpdateDto":{"type":"object","properties":{"kota":{"type":"string","example":"Jakarta"},"jamOperasional":{"type":"string","example":"08:00 - 22:00 WIB"},"lokasi":{"type":"string","example":"Jl. Jenderal Sudirman No. 1, Jakarta"},"image":{"type":"string","example":"/uploads/outlet-jakarta.png"},"googleMapsLink":{"type":"string","example":"https://maps.app.goo.gl/abcdefghijk"},"whatsappUrl":{"type":"string","example":"https://wa.me/6281234567890"},"brandId":{"type":"string","example":"9b62c4d3-5f25-4f8f-83cf-3f49a6b8fd7c"}}},"RequestTeamCreateDto":{"type":"object","properties":{"image":{"type":"string","example":"/uploads/team-john-doe.png"},"nama":{"type":"string","example":"John Doe"},"title":{"type":"string","example":"Chief Executive Officer"},"linkedinUrl":{"type":"string","example":"https://www.linkedin.com/in/john-doe"},"instagramUrl":{"type":"string","example":"https://www.instagram.com/johndoe"}},"required":["image","nama","title","linkedinUrl","instagramUrl"]},"RequestTeamUpdateDto":{"type":"object","properties":{"image":{"type":"string","example":"/uploads/team-john-doe.png"},"nama":{"type":"string","example":"John Doe"},"title":{"type":"string","example":"Chief Executive Officer"},"linkedinUrl":{"type":"string","example":"https://www.linkedin.com/in/john-doe"},"instagramUrl":{"type":"string","example":"https://www.instagram.com/johndoe"}}}}}}